requirements.txt
aioquic==0.9.23
anyio==4.1.0
asgiref==3.7.2
attr==0.3.2
Babel==2.13.1
brotli==1.1.0
brotlicffi==1.1.0.0
ConfigParser==6.0.0
contextlib2==21.6.0
ctypes_snappy==1.03
Cython==3.0.6
dl==0.1.0
docutils==0.20.1
filelock==3.13.1
flask_babel==4.0.0
HTMLParser==0.0.2
httpcore==1.0.2
httpx==0.25.2
ipython==8.18.1
ipywidgets==8.1.1
jnius==1.1.0
kerberos==1.3.1
keyring==24.3.0
mock==5.1.0
Pillow==10.1.0
protobuf==4.25.1
pymongo_auth_aws==1.1.0
pymongocrypt==1.7.0
pyOpenSSL==23.3.0
python-dotenv==1.0.0
railroad==0.5.0
redis==5.0.1
service_identity==23.1.0
simplejson==3.19.2
sniffio==1.3.0
Sphinx==7.2.6
thread==0.1.2
tornado==6.4
trio==0.23.1
trove_classifiers==2023.11.29
urllib3_secure_extra==0.1.0
watchdog==3.0.0
winkerberos==0.9.1
wmi==1.5.1
xmlrpclib==1.0.1
zstandard==0.22.0
